+++ src/types.rs
@@ -1,3 +1,8 @@
+use core::ops::{Add, AddAssign, Div, Mul, Neg, Sub, SubAssign};
+#[cfg(feature = "fixed")]
+use fixed::types::{I16F16, I32F32};
+use num_traits::{FromPrimitive, One, Zero};
+
 /// A type that captures the value range of `i16` but is known to be
 /// fastest on the target platform.
 #[cfg(not(feature = "stdint"))]
@@ -37,3 +42,100 @@ pub type FastUInt8 = u8;
 /// fastest on the target platform.
 #[cfg(feature = "stdint")]
 pub type FastUInt8 = stdint::uint_fast8_t;
+
+/// Collection of numeric traits required for calculations in the Kalman filter.
+/// This trait will be auto-implemented to the majority of relevant types.
+pub trait MatrixDataTypeBase:
+    Add<Self, Output = Self>
+    + AddAssign<Self>
+    + Sub<Self, Output = Self>
+    + SubAssign<Self>
+    + Copy
+    + Zero
+    + One
+    + Mul<Self, Output = Self>
+    + Div<Self, Output = Self>
+    + Neg<Output = Self>
+    + core::iter::Sum<Self>
+    + PartialOrd<Self>
+    + FromPrimitive
+{
+}
+
+impl<T> MatrixDataTypeBase for T where
+    T: Add<T, Output = T>
+        + AddAssign<T>
+        + Sub<T, Output = T>
+        + SubAssign<T>
+        + Copy
+        + Zero
+        + One
+        + Mul<T, Output = T>
+        + Div<T, Output = T>
+        + Neg<Output = T>
+        + core::iter::Sum<T>
+        + PartialOrd<T>
+        + FromPrimitive
+{
+}
+
+/// Trait identifying the data type used by the Kalman Filter matrix operations.
+pub trait MatrixDataType: MatrixDataTypeBase {
+    /// Calculates the reciprocal (inverse) of a number, i.e. `1/self`.
+    fn recip(self) -> Self {
+        Self::one() / self
+    }
+
+    /// Calculates the square root of a number.
+    fn square_root(self) -> Self;
+}
+
+impl MatrixDataType for f32 {
+    /// Calculates the reciprocal (inverse) of a number, i.e. `1/self`.
+    fn recip(self) -> Self {
+        self.recip()
+    }
+
+    /// Calculates the square root of a number.
+    fn square_root(self) -> Self {
+        self.sqrt()
+    }
+}
+
+impl MatrixDataType for f64 {
+    /// Calculates the reciprocal (inverse) of a number, i.e. `1/self`.
+    fn recip(self) -> Self {
+        self.recip()
+    }
+
+    /// Calculates the square root of a number.
+    fn square_root(self) -> Self {
+        self.sqrt()
+    }
+}
+
+#[cfg(feature = "fixed")]
+impl MatrixDataType for I16F16 {
+    /// Calculates the reciprocal (inverse) of a number, i.e. `1/self`.
+    fn recip(self) -> Self {
+        self.recip()
+    }
+
+    /// Calculates the square root of a number.
+    fn square_root(self) -> Self {
+        self.sqrt()
+    }
+}
+
+#[cfg(feature = "fixed")]
+impl MatrixDataType for I32F32 {
+    /// Calculates the reciprocal (inverse) of a number, i.e. `1/self`.
+    fn recip(self) -> Self {
+        self.recip()
+    }
+
+    /// Calculates the square root of a number.
+    fn square_root(self) -> Self {
+        self.sqrt()
+    }
+}